Fix:コース作成に関する処理を共通化
